requirements.txt
openai==0.27.0
pytesseract==0.3.9
pdf2image==1.16.3
PyPDF2==2.11.2
lightrag==0.1.0
python-dotenv==1.0.0
Pillow==9.3.0
